requirements.txt
pytest==8.3.2
requests==2.32.3
allure-pytest==2.13.5
jsonschema==4.23.0
pytest-xdist==3.6.1
